requirements.txt
Flask
langchain
pydantic
python-dotenv
Requests
pytest